+++ src/provider/aws/cloudtrail/cmds/multiRegionTrailEnabled.ts
@@ -29,13 +29,7 @@ export default class MultiRegionTrailEnabled extends AWS {
   global = false
 
   constructor(public params: AWSScannerInterface) {
-    super({
-      profile: params.profile,
-      region: params.region,
-      verbosity: params.verbosity,
-      resourceId: params.resourceId,
-      rule,
-    })
+    super({ ...params, rule })
   }
 
   async audit({ resource, region }: { resource: string; region: string }) {
@@ -100,12 +94,7 @@ export default class MultiRegionTrailEnabled extends AWS {
 }
 
 export const handler = async (args: AWSScannerInterface) => {
-  const scanner = new MultiRegionTrailEnabled({
-    region: args.region,
-    profile: args.profile,
-    resourceId: args.resourceId,
-    verbosity: args.verbosity,
-  })
+  const scanner = new MultiRegionTrailEnabled(args)
 
   await scanner.start()
   scanner.output()