+++ src/provider/aws/ec2/ebs/cmds/snapshotPublic.ts
@@ -23,13 +23,7 @@ export default class PublicSnapshot extends AWS {
   global = false
 
   constructor(public params: AWSScannerInterface) {
-    super({
-      profile: params.profile,
-      resourceId: params.resourceId,
-      region: params.region,
-      verbosity: params.verbosity,
-      rule,
-    })
+    super({ ...params, rule })
   }
 
   async audit({ resource, region }: { resource: string; region: string }) {
@@ -95,12 +89,7 @@ export default class PublicSnapshot extends AWS {
 }
 
 export const handler = async (args: AWSScannerInterface) => {
-  const scanner = new PublicSnapshot({
-    region: args.region,
-    profile: args.profile,
-    resourceId: args.resourceId,
-    verbosity: args.verbosity,
-  })
+  const scanner = new PublicSnapshot(args)
 
   await scanner.start()
   scanner.output()