+++ src/provider/aws/ec2/sg/cmds/publicPermission.ts
@@ -21,13 +21,7 @@ export default class PublicPermission extends AWS {
   service = 'ec2'
 
   constructor(public params: AWSScannerInterface) {
-    super({
-      profile: params.profile,
-      resourceId: params.resourceId,
-      region: params.region,
-      verbosity: params.verbosity,
-      rule,
-    })
+    super({ ...params, rule })
   }
 
   audit = async ({
@@ -95,12 +89,7 @@ export default class PublicPermission extends AWS {
 }
 
 export const handler = async (args: AWSScannerInterface) => {
-  const scanner = await new PublicPermission({
-    region: args.region,
-    profile: args.profile,
-    resourceId: args.resourceId,
-    verbosity: args.verbosity,
-  })
+  const scanner = await new PublicPermission(args)
   await scanner.start()
   scanner.output()
   return scanner.audits