+++ src/provider/aws/elasticache/cmd/encryptionAtRest.ts
@@ -20,8 +20,8 @@ export const desc = `Elasticache instances should have storage encrypted at rest
 
   resourceId: cluster, instance, or shard name
 
-  note: memcached instances do not support encryption therefore they'll always return FAIL
-  note: even standalone instances have a cluster ID
+  note: Memcached instances do not support encryption therefore they'll always return FAIL
+  note: Even standalone instances have a cluster ID
 
 `
 
@@ -31,14 +31,7 @@ export default class AtRestEncryption extends AWS {
   global = false
 
   constructor(public params: AWSScannerInterface) {
-    super({
-      profile: params.profile,
-      resourceId: params.resourceId,
-      region: params.region,
-      verbosity: params.verbosity,
-      keyType: params.keyType,
-      rule,
-    })
+    super({ ...params, rule })
   }
 
   async audit({
@@ -148,13 +141,7 @@ export default class AtRestEncryption extends AWS {
 }
 
 export const handler = async (args: AWSScannerInterface) => {
-  const scanner = new AtRestEncryption({
-    region: args.region,
-    profile: args.profile,
-    resourceId: args.resourceId,
-    keyType: args.keyType,
-    verbosity: args.verbosity,
-  })
+  const scanner = new AtRestEncryption(args)
 
   await scanner.start()
   scanner.output()