+++ src/provider/aws/rds/instances/instanceEncrypted.ts
@@ -30,13 +30,7 @@ export default class PublicInstance extends AWS {
   global = false
 
   constructor(public params: AWSScannerInterface) {
-    super({
-      profile: params.profile,
-      resourceId: params.resourceId,
-      region: params.region,
-      verbosity: params.verbosity,
-      rule,
-    })
+    super({ ...params, rule })
     this.keyType = params.keyType
   }
 
@@ -97,13 +91,7 @@ export default class PublicInstance extends AWS {
 }
 
 export const handler = async (args: AWSScannerInterface) => {
-  const scanner = new PublicInstance({
-    region: args.region,
-    profile: args.profile,
-    resourceId: args.resourceId,
-    keyType: args.keyType,
-    verbosity: args.verbosity,
-  })
+  const scanner = new PublicInstance(args)
 
   await scanner.start()
   scanner.output()