+++ src/provider/aws/rds/instances/publicInstance.ts
@@ -26,13 +26,7 @@ export default class PublicInstance extends AWS {
   global = false
 
   constructor(public params: AWSScannerInterface) {
-    super({
-      profile: params.profile,
-      resourceId: params.resourceId,
-      region: params.region,
-      verbosity: params.verbosity,
-      rule,
-    })
+    super({ ...params, rule })
   }
 
   async audit({
@@ -88,12 +82,7 @@ export default class PublicInstance extends AWS {
 }
 
 export const handler = async (args: AWSScannerInterface) => {
-  const scanner = new PublicInstance({
-    region: args.region,
-    profile: args.profile,
-    resourceId: args.resourceId,
-    verbosity: args.verbosity,
-  })
+  const scanner = new PublicInstance(args)
 
   await scanner.start()
   scanner.output()